Modularize data handling utilities for TabNet.
Refactored data handling components into separate modules for improved clarity and maintainability. Added distinct classes for datasets, data loaders, and related utility functions in separate files, aligning with the single-responsibility principle.
